Implemented Player Skin & Fixed Ability Collider
The player skin is now the "Default" skin and this was achieved through setting the "Player" Objects Skinned Mesh Renderer's Mesh to the "default" mesh. And setting the Skinned Mesh Renderer's Materials to the "default's" materials.
The Ability seemed to be too accurate so I made the sollision sphere bigger in order to be more forgiving